feat: service container 분리, event packet 전송

+++ PenguinParty/Dto/Player.cs
@@ -4,5 +4,5 @@ public class Player(ulong userId)
 {
     public readonly ulong UserId = userId;
     public List<Card> Cards { get; } = [];
-    public byte Score;
+    public int Score;
 }
